[FIND MEASURES] tracking of measure_sids - code improvements
